게시글작성,읽기 조금바꿈

+++ WebContent/jsp_project/recipes/content.jsp
@@ -1,25 +1,42 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
-<%@ page import = "DBBean.BoardDataBean" %>
+<%@ page import = "DBBean.foodingBean" %>
 <%@ page import = "DBBean.BoardDataBean" %>
 <%@ page import = "java.text.SimpleDateFormat" %>
-<%@ include file="color.jsp"%>
-
+<!DOCTYPE html>
 <html>
 <head>
+<meta name="viewport" content="width=device-width, initial-scale=1">
+
+<link rel="stylesheet" href="../css/common.css">
+<style>
+
+
+
+
+
+</style>
 <title>게시판</title>
-<link href="style.css" rel="stylesheet" type="text/css">
+
+<script type="text/javascript">
+
+</script>
+
 </head>
-<body bgcolor="<%=bodyback_c%>">
+<body>
+<%@include file="../general_included/topbar.jsp"%>
+
+<div id="maindiv">
 <%
 	int num = Integer.parseInt(request.getParameter("num"));
    String pageNum = request.getParameter("pageNum");
 
    SimpleDateFormat sdf = 
         new SimpleDateFormat("yyyy-MM-dd HH:mm");
+   foodingBean foodingbean = new foodingBean();
 
    try{
-      BoardDataBean dbPro = BoardDataBean.getInstance(); 
+	   foodingBean dbPro = foodingBean.getInstance(); 
       BoardDataBean article =  dbPro.getArticle(num);
 %>
 
@@ -28,48 +45,66 @@
 <form>
 <table>  
   <tr height="30">
-    <td align="center" width="125" bgcolor="<%=value_c%>">글번호</td>
+    <td align="center" width="125" >글번호</td>
     <td align="center" width="125" align="center">
 	     <%=article.getNum()%></td>
-    <td align="center" width="125" bgcolor="<%=value_c%>">조회수</td>
+    <td align="center" width="125" >조회수</td>
     <td align="center" width="125" align="center">
 	     <%=article.getReadcount()%></td>
   </tr>
   <tr height="30">
-    <td align="center" width="125" bgcolor="<%=value_c%>">작성자</td>
+    <td align="center" width="125" >작성자</td>
     <td align="center" width="125" align="center">
-	     <%=article.getWriterid()%></td>
-    <td align="center" width="125" bgcolor="<%=value_c%>" >작성일</td>
+	     <%=foodingbean.findnkname(article.getWriterid())%></td>
+    <td align="center" width="125"  >작성일</td>
     <td align="center" width="125" align="center">
 	     <%= sdf.format(article.getReg_date())%></td>
   </tr>
   <tr height="30">
-    <td align="center" width="125" bgcolor="<%=value_c%>">글제목</td>
+    <td align="center" width="125" >글제목</td>
     <td align="center" width="375" align="center" colspan="3">
 	     <%=article.getTitle()%></td>
   </tr>
   <tr>
-    <td align="center" width="125" bgcolor="<%=value_c%>">글내용</td>
+    <td align="center" width="125" >글내용</td>
     <td align="left" width="375" colspan="3">
            <pre><%=article.getContent()%></pre></td>
   </tr>
   <tr height="30">      
-    <td colspan="4" bgcolor="<%=value_c%>" align="right" > 
-	  <input type="button" value="글수정" 
-       onclick="document.location.href='updateForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
-	   &nbsp;&nbsp;&nbsp;&nbsp;
-	  <input type="button" value="글삭제" 
-       onclick="document.location.href='deleteForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
+    <td colspan="4"  align="right" > 
+    <%
+    if(article.getWriterid().equals((String)session.getAttribute("idlogin"))){
+        %>
+        	<input type="button" value="글수정" 
+           		onclick="document.location.href='updateForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
+    	   		&nbsp;&nbsp;&nbsp;&nbsp;
+    	  <input type="button" value="글삭제" 
+           		onclick="document.location.href='deleteForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
+        <%
+        }
+        %>
+	  
 	   &nbsp;&nbsp;&nbsp;&nbsp;
 	   &nbsp;&nbsp;&nbsp;&nbsp;
        <input type="button" value="글목록" 
        onclick="document.location.href='list.jsp?pageNum=<%=pageNum%>'">
     </td>
   </tr>
-</table>    
+</table>
 <%
  }catch(Exception e){} 
  %>
 </form>      
+</div>
+
+
+
+<div id="footer" align="right" style="color:#cccccc; font-size:12px;">
+<pre>
+Create by FOODING
+고객문의 1544-XXXX
+JSP Project 2019 2A03</pre>
+</div>
+
 </body>
 </html>
